thirteen clovers!
And he carried me away in the Spirit to a mountain great and high, and showed me the holy city of Jerusalem coming down out of heaven from God, shining with the glory of God. Its radiance was like a most precious jewel, like a jasper, as clear as crystal. The city had a great and high wall with twelve gates inscribed with the names of the twelve tribes of Israel, and twelve angels at the gates. There were three gates on the east, three on the north, three on the south, and three on the west. The wall of the city had twelve foundations bearing the names of the twelve apostles of the Lamb. 
Revelation 21